refactor - user: In/Ex 재분류
